refactor: WinningNumbersTest 분리
