+++ src/test/java/lotto/model/LottoMatcherTest.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.model;
 
 import lotto.model.Lotto;
 import lotto.model.LottoMatcher;
@@ -72,41 +72,5 @@ class LottoMatcherTest {
             new LottoMatcher(Arrays.asList(1, 2, 3, 4, 5, 6), bonusNumber);
         }).isInstanceOf(IllegalArgumentException.class)
                 .hasMessageContaining("보너스 볼 번호가 1~45 범위 내에 해당하지 않습니다.");
-    }
-
-    @Test
-    @DisplayName("지난 주 당첨 번호 개수가 6개 미만인 경우 예외 처리")
-    void validateNumberOfWinningNumbersUnder6Test() {
-        assertThatThrownBy(() -> {
-            new LottoMatcher(Arrays.asList(1, 2, 3, 4, 5), 45);
-        }).isInstanceOf(IllegalArgumentException.class)
-                .hasMessageContaining("지난 주 당첨 번호 개수는 6개로 입력해주세요.");
-    }
-
-    @Test
-    @DisplayName("지난 주 당첨 번호 개수가 6개 초과인 경우 예외 처리")
-    void validateNumberOfWinningNumbersOver6Test() {
-        assertThatThrownBy(() -> {
-            new LottoMatcher(Arrays.asList(1, 2, 3, 4, 5, 6, 7), 45);
-        }).isInstanceOf(IllegalArgumentException.class)
-                .hasMessageContaining("지난 주 당첨 번호 개수는 6개로 입력해주세요.");
-    }
-
-    @Test
-    @DisplayName("지난 주 당첨 번호의 범위가 1~45가 아닌 경우 예외 처리")
-    void validateRangeWinningNumbersTest() {
-        assertThatThrownBy(() -> {
-            new LottoMatcher(Arrays.asList(1, 2, 3, 4, 55, 6), 45);
-        }).isInstanceOf(IllegalArgumentException.class)
-                .hasMessageContaining("당첨 번호의 범위는 1 ~ 45 사이로 입력해주세요.");
-    }
-
-    @Test
-    @DisplayName("지난 주 당첨 번호 중 중복된 값이 입력될 경우 예외 처리")
-    void validateDuplicationWinningNumbersTest() {
-        assertThatThrownBy(() -> {
-            new LottoMatcher(Arrays.asList(1, 2, 3, 3, 4, 6), 45);
-        }).isInstanceOf(IllegalArgumentException.class)
-                .hasMessageContaining("당첨 번호에 중복이 존재합니다.");
     }
 }
